requirements.txt
python-telegram-bot==21.9
gspread>=5.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
pandas>=2.0.0
schedule>=1.2.0
pytz>=2023.3
python-dateutil>=2.8.2
Pillow>=10.0.0
